add 5dim "core" compatibility
When 5dim's core mod is enabled, it alters the vanilla pipes to be in a different group and subgroup. This makes high-pressure pipes unable to be an upgrade option for the vanilla pipes. Also, it means the high-pressure pipes are not in the same group in the GUI as the vanilla pipes.

This rectifies both of those issues by changing the high-pressure pipes subgroup and ordering to better align with 5dim's way of doing things, if 5dim's core mod is enabled.